pyproject.toml
[project]
name = "message-reader"
version = "0.0.1"
description = "A discord application that stores messages"
authors = ["Thibault Latxague <thibault.latxague@gmail.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
]
